Refactor custom createBlock to utils
